Finalize app layout; fix an issue with splitter index
