+++ src/components/Canvas/index.ts
@@ -1,0 +1,1 @@
+export * from './Canvas';
