--- src/components/Draggable/index.ts
@@ -1,1 +1,0 @@
-export * from './Draggable';
