+++ src/components/Editor/index.ts
@@ -1,0 +1,1 @@
+export * from './Editor';
