--- src/components/PanelLayout/index.ts
@@ -1,1 +1,0 @@
-export * from './PanelLayout';
