--- src/components/Sidebar/index.ts
@@ -1,1 +1,0 @@
-export * from './Sidebar';
