+++ src/components/Toolbar.tsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
 
-const Wrapper = styled.div`
+const ToolbarWrapper = styled.div`
   flex: 1 1 auto;
   background: lightskyblue;
 `;
 
-export class Sidebar extends Component {
+export class Toolbar extends Component {
   render() {
-    return <Wrapper>Sidebar</Wrapper>;
+    return <ToolbarWrapper>Toolbar</ToolbarWrapper>;
   }
 }
