+++ src/components/Workbench.tsx
@@ -1,13 +1,12 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
-import { Canvas } from '../Canvas/Canvas';
-import { Editor } from '../Editor/Editor';
-import { PanelLayout } from '../PanelLayout';
-import { Panel } from '../PanelLayout/Panel';
-import { Splitter } from '../PanelLayout/Splitter';
-import { Sidebar } from '../Sidebar';
+import { Canvas } from './Canvas';
+import { Editor } from './Editor';
+import { Panel, PanelLayout, Splitter } from './shared/PanelLayout';
+import { Sidebar } from './Sidebar';
+import { Toolbar } from './Toolbar';
 
-const Wrapper = styled.div`
+const WorkbenchWrapper = styled.div`
   flex: 1 1 auto;
   display: flex;
   position: relative;
@@ -16,10 +15,13 @@ const Wrapper = styled.div`
 export class Workbench extends Component {
   render() {
     return (
-      <Wrapper style={{}}>
+      <WorkbenchWrapper style={{}}>
         <PanelLayout>
-          <Panel initSize={400} minSize={160} fixed={true}>
+          <Panel initSize={48} minSize={48} maxSize={48} fixed={true}>
             <Sidebar />
+          </Panel>
+          <Panel initSize={360} minSize={160} fixed={true}>
+            <Toolbar />
           </Panel>
           <Splitter />
           <Panel minSize={480}>
@@ -34,7 +36,7 @@ export class Workbench extends Component {
             </PanelLayout>
           </Panel>
         </PanelLayout>
-      </Wrapper>
+      </WorkbenchWrapper>
     );
   }
 }
