--- src/components/Workbench/index.ts
@@ -1,1 +1,0 @@
-export * from './Workbench';
