+++ src/components/shared/PanelLayout/index.ts
@@ -1,0 +1,3 @@
+export * from './PanelLayout';
+export * from './Panel';
+export * from './Splitter';
